refactor: auth rename to member #75
